Remove count of X and O from Board
The count of X and Os within the Board was only used to check if the board was full when checking for a draw. However, there is already a function to check for empty cells which can be used instead.

Also changes Board::check() to return the current state of the Board, moving any printing of the draw result out of that function and into Main along with the winning result.