Add HardAIPlayer
HardAIPlayer::makeAIMove is implemented using the Minimax algorithm.

Within the algorithm it is required to switch between which player will be taking a move. To keep track of this, an opponent field is added to the AIPlayer.

As processing the algorithm requires making moves on the board, to ensure the main board immutable a new function is added to Board to create a copy.